Inital Commit - Import
Basic script created for importing the data we recive and pushing it to
a matlab structure, then we can pull it out and push it back into an
excel file.

The paths for the pictures are also found, so we can pull the photos to
manage now.

Not very user friendly.