Added Spline, Video, and Calculations
Created a spline to reduce noise. Not perfect but it does give some idea
of the general shape.

Also added video creation, with information overlays. Currently no
images are written like this but it would be simple to add.

Also added in instintainous slew and a format to add in any other
calculations if we want to.